Add support for Yongfeng tones
The yue and hak files both use tone numbers, while cmn uses the Chao tone scale. As a southern Chinese language, the Yongfeng tone system is etymologically similar to Hakka, but its sandhi rules are more complex. In order to successfully represent allotones, I'm attempting to use a mixed system, with single digits for phonemic tones and Chao numbers for the sandhi realizations.